Async API is intoduced for Route4MeManager and Route4MeManagerV5.
Mix of sync/async context is removed for async flow.
IDs are switched from 32 bit integer to 64 bit integer (both signed and unsigned).
Async API for FastProcessing (JSON) is introduced.
Unnesessary internal state is removed from FastProcessing.
ArrayList is removed from geocoding API.
Internal sleeps and spin waits are removed.
Naming is updated to follow CamelCase.
Internal code smells are removed.
NewtonJson usage is fixed.

+++ route4me-csharp-sdk/Route4MeSDKLibrary/DataTypes/AddressBookContact.cs
@@ -35,7 +35,7 @@ namespace Route4MeSDK.DataTypes
         ///     Unique ID of the contact.
         /// </summary>
         [DataMember(Name = "address_id", EmitDefaultValue = false)]
-        public int? AddressId { get; set; }
+        public long? AddressId { get; set; }
 
         /// <summary>
         ///     The geographic address of the contact.
